Remove HeteroGraphWrapper and move model to entity_classify.py
