+++ examples/pytorch/mock_sparse/rgcn-hetero/entity_classify.py
@@ -4,14 +4,64 @@ Reference Code: https://github.com/tkipf/relational-gcn
 """
 import argparse
 import numpy as np
-import time
 import torch as th
 import torch.nn as nn
 import torch.nn.functional as F
 
 from dgl.data.rdf import AIFBDataset, MUTAGDataset, BGSDataset, AMDataset
-from model import EntityClassify
-from hetero_graph_wrapper import HeteroGraphWrapper
+
+from dgl.mock_sparse import create_from_coo
+
+
+class RelGraphConvLayer(nn.Module):
+    def __init__(self, in_feat, out_feat, rel_names):
+        super(RelGraphConvLayer, self).__init__()
+        self.in_feat = in_feat
+        self.out_feat = out_feat
+        self.rel_names = rel_names
+        self.weight = {
+            rel: nn.Parameter(th.Tensor(in_feat, out_feat))
+            for rel in self.rel_names
+        }
+        for w in self.weight.values():
+            nn.init.xavier_uniform_(w, gain=nn.init.calculate_gain("relu"))
+
+    def forward(self, adjs, x_dict):
+        h_dict = {ntype: 0 for ntype in x_dict.keys()}
+        for stype, etype, dtype in adjs.keys():
+            h = x_dict[stype] @ self.weight[etype]  # dense mm
+            h_dict[dtype] += adjs[(stype, etype, dtype)] @ h
+        h_dict = {ntype: th.relu(h) for ntype, h in h_dict.items()}
+        return h_dict
+
+
+class EntityClassify(nn.Module):
+    def __init__(self, adjs, ntype_to_num_nodes, h_dim, out_dim):
+        super(EntityClassify, self).__init__()
+        self.adjs = adjs
+        self.h_dim = h_dim
+        self.out_dim = out_dim
+        self.rel_names = list(set(etype for _, etype, _ in adjs.keys()))
+
+        self.embeds = nn.ParameterDict()
+        for ntype, num_nodes in ntype_to_num_nodes.items():
+            embed = nn.Parameter(th.Tensor(num_nodes, h_dim))
+            nn.init.xavier_uniform_(embed, gain=nn.init.calculate_gain("relu"))
+            self.embeds[ntype] = embed
+
+        self.layers = nn.ModuleList()
+        self.layers.append(
+            RelGraphConvLayer(self.h_dim, self.h_dim, self.rel_names)
+        )
+        self.layers.append(
+            RelGraphConvLayer(self.h_dim, self.out_dim, self.rel_names)
+        )
+
+    def forward(self):
+        h = self.embeds
+        for layer in self.layers:
+            h = layer(self.adjs, h)
+        return h
 
 
 def main(args):
@@ -35,63 +85,36 @@ def main(args):
     train_idx = th.nonzero(train_mask, as_tuple=False).squeeze()
     test_idx = th.nonzero(test_mask, as_tuple=False).squeeze()
     labels = g.nodes[category].data.pop("labels")
-    category_id = len(g.ntypes)
-    for i, ntype in enumerate(g.ntypes):
-        if ntype == category:
-            category_id = i
+    val_idx = train_idx
 
-    # split dataset into train, validate, test
-    if args.validation:
-        val_idx = train_idx[: len(train_idx) // 5]
-        train_idx = train_idx[len(train_idx) // 5 :]
-    else:
-        val_idx = train_idx
+    def create_adjs(g):
+        adjs = {}
+        for rel in g.canonical_etypes:
+            stype, _, dtype = rel
+            row, col = g.edges(etype=rel)
+            adjs[rel] = create_from_coo(
+                col,
+                row,
+                shape=(g.number_of_nodes(dtype), g.number_of_nodes(stype)),
+            )
+        return adjs
 
-    # check cuda
-    use_cuda = args.gpu >= 0 and th.cuda.is_available()
-    if use_cuda:
-        th.cuda.set_device(args.gpu)
-        g = g.to("cuda:%d" % args.gpu)
-        labels = labels.cuda()
-        train_idx = train_idx.cuda()
-        test_idx = test_idx.cuda()
+    # Dict: canonical_etype -> SparseMatrix
+    adjs = create_adjs(g)
+    ntype_to_num_nodes = {ntype: g.number_of_nodes(ntype) for ntype in g.ntypes}
+    model = EntityClassify(adjs, ntype_to_num_nodes, 16, num_classes)
 
-    g = HeteroGraphWrapper(g)
-    # create model
-    model = EntityClassify(
-        g,
-        args.n_hidden,
-        num_classes,
-        num_bases=args.n_bases,
-        num_hidden_layers=args.n_layers - 2,
-        dropout=args.dropout,
-        use_self_loop=args.use_self_loop,
-    )
+    optimizer = th.optim.Adam(model.parameters(), lr=1e-2)
 
-    if use_cuda:
-        model.cuda()
-
-    # optimizer
-    optimizer = th.optim.Adam(
-        model.parameters(), lr=args.lr, weight_decay=args.l2norm
-    )
-
-    # training loop
     print("start training...")
-    dur = []
     model.train()
-    for epoch in range(args.n_epochs):
+    for epoch in range(50):
         optimizer.zero_grad()
-        if epoch > 5:
-            t0 = time.time()
         logits = model()[category]
         loss = F.cross_entropy(logits[train_idx], labels[train_idx])
         loss.backward()
         optimizer.step()
-        t1 = time.time()
 
-        if epoch > 5:
-            dur.append(t1 - t0)
         train_acc = th.sum(
             logits[train_idx].argmax(dim=1) == labels[train_idx]
         ).item() / len(train_idx)
@@ -100,18 +123,15 @@ def main(args):
             logits[val_idx].argmax(dim=1) == labels[val_idx]
         ).item() / len(val_idx)
         print(
-            "Epoch {:05d} | Train Acc: {:.4f} | Train Loss: {:.4f} | Valid Acc: {:.4f} | Valid loss: {:.4f} | Time: {:.4f}".format(
+            "Epoch {:05d} | Train Acc: {:.4f} | Train Loss: {:.4f} | Valid Acc: {:.4f} | Valid loss: {:.4f}".format(
                 epoch,
                 train_acc,
                 loss.item(),
                 val_acc,
                 val_loss.item(),
-                np.average(dur),
             )
         )
     print()
-    if args.model_path is not None:
-        th.save(model.state_dict(), args.model_path)
 
     model.eval()
     logits = model.forward()[category]
@@ -130,46 +150,12 @@ def main(args):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="RGCN")
     parser.add_argument(
-        "--dropout", type=float, default=0, help="dropout probability"
+        "-d",
+        "--dataset",
+        type=str,
+        required=True,
+        help="Dataset name ('aifb', 'mutag', 'bgs', 'am').",
     )
-    parser.add_argument(
-        "--n-hidden", type=int, default=16, help="number of hidden units"
-    )
-    parser.add_argument("--gpu", type=int, default=-1, help="gpu")
-    parser.add_argument("--lr", type=float, default=1e-2, help="learning rate")
-    parser.add_argument(
-        "--n-bases",
-        type=int,
-        default=-1,
-        help="number of filter weight matrices, default: -1 [use all]",
-    )
-    parser.add_argument(
-        "--n-layers", type=int, default=2, help="number of propagation rounds"
-    )
-    parser.add_argument(
-        "-e",
-        "--n-epochs",
-        type=int,
-        default=50,
-        help="number of training epochs",
-    )
-    parser.add_argument(
-        "-d", "--dataset", type=str, required=True, help="dataset to use"
-    )
-    parser.add_argument(
-        "--model_path", type=str, default=None, help="path for save the model"
-    )
-    parser.add_argument("--l2norm", type=float, default=0, help="l2 norm coef")
-    parser.add_argument(
-        "--use-self-loop",
-        default=False,
-        action="store_true",
-        help="include self feature as a special relation",
-    )
-    fp = parser.add_mutually_exclusive_group(required=False)
-    fp.add_argument("--validation", dest="validation", action="store_true")
-    fp.add_argument("--testing", dest="validation", action="store_false")
-    parser.set_defaults(validation=True)
 
     args = parser.parse_args()
     print(args)
